feat: create project-stats component #85
