+++ components/Project.vue
@@ -1,17 +1,5 @@
 <script lang="ts">
-import dayjs from 'dayjs'
-import localizedFormat from 'dayjs/plugin/localizedFormat'
-import relativeTime from 'dayjs/plugin/relativeTime'
-import en from 'dayjs/locale/en-gb'
-import fr from 'dayjs/locale/fr'
-import es from 'dayjs/locale/es'
 import type { Project } from '~/libs/types'
-
-dayjs.extend(localizedFormat)
-dayjs.extend(relativeTime)
-
-// Force to import and keep locales
-const _daysjsLocale = { en, fr, es }
 
 export default defineNuxtComponent({
   name: 'Project',
@@ -30,17 +18,6 @@ export default defineNuxtComponent({
     overpassUrl(): string {
       return `${this.apiUrl}/projects/${this.project.id}/overpasslike/`
     },
-    locale(): string {
-      return this.$i18n.locale
-    },
-    lastUpdate(): string {
-      return dayjs(this.project.date_last_update)
-        .locale(this.$i18n.locale)
-        .fromNow()
-    },
-    toBeValidated(): string | undefined {
-      return this.project.to_be_validated?.toLocaleString(this.$i18n.locale)
-    },
   },
 })
 </script>
@@ -51,32 +28,7 @@ export default defineNuxtComponent({
       <project-light :project="project" title-tag="h3" />
     </template>
     <div>
-      <el-row id="stats">
-        <el-col :span="12">
-          <div class="el-statistic">
-            <div class="el-statistic__head">
-              {{ $t('project.lastUpdate') }}
-            </div>
-            <div class="el-statistic__content">
-              <span class="el-statistic__number">
-                {{ lastUpdate }}
-              </span>
-            </div>
-          </div>
-        </el-col>
-        <el-col :span="12">
-          <div class="el-statistic">
-            <div class="el-statistic__head">
-              {{ $t('project.toBeValidated') }}
-            </div>
-            <div class="el-statistic__content">
-              <span class="el-statistic__number">
-                {{ toBeValidated }}
-              </span>
-            </div>
-          </div>
-        </el-col>
-      </el-row>
+      <project-stats :project="project" />
       <el-divider border-style="dotted" />
       <el-row>
         <UserGroups :user-groups="Object.values(project.user_groups)" />
@@ -110,32 +62,3 @@ export default defineNuxtComponent({
     <p>{{ $t('app.project.join') }}</p>
   </el-card>
 </template>
-
-<style scoped>
-#stats .el-col {
-  text-align: center;
-}
-
-#stats .el-statistic {
-  --el-statistic-title-font-weight: 400;
-  --el-statistic-title-font-size: var(--el-font-size-extra-small);
-  --el-statistic-title-color: var(--el-text-color-regular);
-  --el-statistic-content-font-weight: 400;
-  --el-statistic-content-font-size: var(--el-font-size-extra-large);
-  --el-statistic-content-color: var(--el-text-color-primary);
-}
-
-#stats .el-statistic__head {
-  font-weight: var(--el-statistic-title-font-weight);
-  font-size: var(--el-statistic-title-font-size);
-  color: var(--el-statistic-title-color);
-  line-height: 20px;
-  margin-bottom: 4px;
-}
-
-#stats .el-statistic__content {
-  font-weight: var(--el-statistic-content-font-weight);
-  font-size: var(--el-statistic-content-font-size);
-  color: var(--el-statistic-content-color);
-}
-</style>
